+++ crates/oxc_linter/src/rules/vue/no_expose_after_await.rs
@@ -58,7 +58,7 @@ declare_oxc_lint!(
     NoExposeAfterAwait,
     vue,
     correctness,
-    version = "next",
+    version = "1.67.0",
 );
 
 impl Rule for NoExposeAfterAwait {