+++ crates/oxc_parser/src/js/class.rs
@@ -73,7 +73,10 @@ impl<'a, C: Config> ParserImpl<'a, C> {
             start_span = d.span.start;
         }
 
-        let id = if self.cur_kind().is_binding_identifier() && !self.at(Kind::Implements) {
+        let id = if self.cur_kind().is_binding_identifier()
+            && !(self.at(Kind::Implements)
+                && self.lexer.peek_token().kind().is_identifier_or_keyword())
+        {
             Some(self.parse_binding_identifier())
         } else {
             None
@@ -405,6 +408,14 @@ impl<'a, C: Config> ParserImpl<'a, C> {
             true,
             diagnostics::accessor_modifier,
         );
+        if definite
+            && ((self.ctx.has_ambient() && !modifiers.contains(ModifierKind::Declare))
+                || r#type.is_abstract())
+        {
+            self.error(diagnostics::definite_assignment_assertion_not_permitted(
+                self.end_span(span),
+            ));
+        }
         self.ast.class_element_accessor_property(
             self.end_span(span),
             r#type,
@@ -677,6 +688,11 @@ impl<'a, C: Config> ParserImpl<'a, C> {
                 initializer.span(),
             ));
         }
+        if definite && (self.ctx.has_ambient() || r#abstract) {
+            self.error(diagnostics::definite_assignment_assertion_not_permitted(
+                self.end_span(span),
+            ));
+        }
         self.ast.class_element_property_definition(
             self.end_span(span),
             r#type,
@@ -697,22 +713,35 @@ impl<'a, C: Config> ParserImpl<'a, C> {
 
     #[cold]
     pub(crate) fn check_getter(&mut self, function: &Function<'a>) {
-        if !function.params.items.is_empty() {
+        if let Some(type_parameters) = &function.type_parameters {
+            self.error(diagnostics::accessor_cannot_have_type_parameters(type_parameters.span));
+        } else if !function.params.items.is_empty() {
             self.error(diagnostics::getter_parameters(function.params.span));
         }
     }
 
     #[cold]
     pub(crate) fn check_setter(&mut self, function: &Function<'a>) {
-        if let Some(rest) = &function.params.rest {
-            self.error(diagnostics::setter_with_rest_parameter(rest.span));
+        if let Some(type_parameters) = &function.type_parameters {
+            self.error(diagnostics::accessor_cannot_have_type_parameters(type_parameters.span));
         } else if function.params.parameters_count() != 1 {
             self.error(diagnostics::setter_with_parameters(
                 function.params.span,
                 function.params.parameters_count(),
             ));
-        } else if self.is_ts && function.params.items.first().unwrap().initializer.is_some() {
-            self.error(diagnostics::setter_with_initializer(function.params.span));
+        } else if let Some(rest) = &function.params.rest {
+            self.error(diagnostics::setter_with_rest_parameter(rest.span));
+        } else if self.is_ts {
+            let param = function.params.items.first().unwrap();
+            if let Some(return_type) = &function.return_type {
+                self.error(diagnostics::a_set_accessor_cannot_have_a_return_type_annotation(
+                    return_type.span(),
+                ));
+            } else if param.optional {
+                self.error(diagnostics::setter_with_optional_parameter(param.span));
+            } else if param.initializer.is_some() {
+                self.error(diagnostics::setter_with_initializer(function.params.span));
+            }
         }
     }